test(statistic): 정의를 V066 적용 후 상태로 맞추고 name 길이 회귀를 추가
앞 커밋이 드러낸 오버플로의 실제 수정은 WordOnlineDatabase#75의
V066_20260824__widen_statistic_update_time_intervals.sql이다. H2는 마이그레이션을
돌리지 않으므로, 테스트 스키마는 마이그레이션 체인의 head 상태를 반영한다:
min/max는 BIGINT, name은 varchar(31), mean은 double precision.

단정을 완화하지 않았다. 오버플로 테스트는 그대로 4초 간격을 요구하고, 그 요구가
성립하려면 V066이 대상 데이터베이스에 먼저 적용되어야 한다는 사실을 스키마 파일
주석과 테스트 javadoc에 명시했다.

varchar(31)이 실제 GameSystem 이름을 자르지 않는지 확인하는 회귀를 추가했다.
가장 긴 이름은 GameObjectStateInitialSystem으로 28자다.

+++ src/test/resources/schema-h2.sql
@@ -142,16 +142,19 @@ CREATE TABLE statistic_game_magics (
 CREATE INDEX idx_statistic_game_magic_user_id_statistic_game_id
     ON statistic_game_magics(user_id, statistic_game_id);
 
--- Mirrors database/migration/V000_20260406__init_tables.sql line 508. Every column type,
--- length and nullability below is copied from that definition, not inferred from the
--- INSERT in StatisticRepository. The inferred version claimed BIGINT intervals and a
--- VARCHAR(255) name; production has INTEGER and varchar(31).
+-- Copied from the production definition, not inferred from the INSERT in
+-- StatisticRepository: database/migration/V000_20260406__init_tables.sql line 508, as
+-- amended by V066_20260824__widen_statistic_update_time_intervals.sql. V000 declared the
+-- interval columns INTEGER, which caps them at ~2.147 seconds; V066 widens them to
+-- BIGINT. This file must not run ahead of that migration -- see WordOnlineDatabase#75.
+-- name is varchar(31) and mean_interval_ns is double precision in V000; the earlier
+-- inferred definition guessed VARCHAR(255) and REAL.
 CREATE TABLE statistic_update_time (
     id BIGSERIAL PRIMARY KEY,
     statistic_game_id BIGINT REFERENCES statistic_games(id) ON DELETE CASCADE,
     name VARCHAR(31),
-    min_interval_ns INTEGER,
-    max_interval_ns INTEGER,
+    min_interval_ns BIGINT,
+    max_interval_ns BIGINT,
     mean_interval_ns DOUBLE PRECISION
 );
